+++ sei-ibc-go/modules/apps/27-interchain-accounts/host/keeper/handshake.go
@@ -105,24 +105,19 @@ func (k Keeper) OnChanCloseConfirm(
 // validateControllerPortParams asserts the provided connection sequence and counterparty connection sequence
 // match that of the associated connection stored in state
 func (k Keeper) validateControllerPortParams(ctx sdk.Context, channelID, portID string, connectionSeq, counterpartyConnectionSeq uint64) error {
-	channel, found := k.channelKeeper.GetChannel(ctx, portID, channelID)
-	if !found {
-		return sdkerrors.Wrapf(channeltypes.ErrChannelNotFound, "port ID %s channel ID %s", portID, channelID)
-	}
-
-	counterpartyHops, found := k.channelKeeper.CounterpartyHops(ctx, channel)
-	if !found {
-		return sdkerrors.Wrap(connectiontypes.ErrConnectionNotFound, channel.ConnectionHops[0])
-	}
-
-	connSeq, err := connectiontypes.ParseConnectionSequence(channel.ConnectionHops[0])
+	connectionID, connection, err := k.channelKeeper.GetChannelConnection(ctx, portID, channelID)
 	if err != nil {
-		return sdkerrors.Wrapf(err, "failed to parse connection sequence %s", channel.ConnectionHops[0])
+		return err
 	}
 
-	counterpartyConnSeq, err := connectiontypes.ParseConnectionSequence(counterpartyHops[0])
+	connSeq, err := connectiontypes.ParseConnectionSequence(connectionID)
 	if err != nil {
-		return sdkerrors.Wrapf(err, "failed to parse counterparty connection sequence %s", counterpartyHops[0])
+		return sdkerrors.Wrapf(err, "failed to parse connection sequence %s", connectionID)
+	}
+
+	counterpartyConnSeq, err := connectiontypes.ParseConnectionSequence(connection.GetCounterparty().GetConnectionID())
+	if err != nil {
+		return sdkerrors.Wrapf(err, "failed to parse counterparty connection sequence %s", connection.GetCounterparty().GetConnectionID())
 	}
 
 	if connSeq != connectionSeq {