+++ sei-ibc-go/modules/core/04-channel/keeper/handshake.go
@@ -5,33 +5,13 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
+
 	connectiontypes "github.com/cosmos/ibc-go/v3/modules/core/03-connection/types"
 	"github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
 	porttypes "github.com/cosmos/ibc-go/v3/modules/core/05-port/types"
 	host "github.com/cosmos/ibc-go/v3/modules/core/24-host"
 	"github.com/cosmos/ibc-go/v3/modules/core/exported"
 )
-
-// CounterpartyHops returns the connection hops of the counterparty channel.
-// The counterparty hops are stored in the inverse order as the channel's.
-// NOTE: Since connectionHops only supports single connection channels for now,
-// this function requires that connection hops only contain a single connection id
-func (k Keeper) CounterpartyHops(ctx sdk.Context, ch types.Channel) ([]string, bool) {
-	// Return empty array if connection hops is more than one
-	// ConnectionHops length should be verified earlier
-	if len(ch.ConnectionHops) != 1 {
-		return []string{}, false
-	}
-	counterpartyHops := make([]string, 1)
-	hop := ch.ConnectionHops[0]
-	conn, found := k.connectionKeeper.GetConnection(ctx, hop)
-	if !found {
-		return []string{}, false
-	}
-
-	counterpartyHops[0] = conn.GetCounterparty().GetConnectionID()
-	return counterpartyHops, true
-}
 
 // ChanOpenInit is called by a module to initiate a channel opening handshake with
 // a module on another chain. The counterparty channel identifier is validated to be
@@ -127,6 +107,11 @@ func (k Keeper) ChanOpenTry(
 
 	channelID := previousChannelID
 
+	// connection hops only supports a single connection
+	if len(connectionHops) != 1 {
+		return "", nil, sdkerrors.Wrapf(types.ErrTooManyConnectionHops, "expected 1, got %d", len(connectionHops))
+	}
+
 	// empty channel identifier indicates continuing a previous channel handshake
 	if previousChannelID != "" {
 		// channel identifier and connection hop length checked on msg.ValidateBasic()
@@ -139,7 +124,7 @@ func (k Keeper) ChanOpenTry(
 		if !(previousChannel.Ordering == order &&
 			previousChannel.Counterparty.PortId == counterparty.PortId &&
 			previousChannel.Counterparty.ChannelId == "" &&
-			previousChannel.ConnectionHops[0] == connectionHops[0] &&
+			previousChannel.ConnectionHops[0] == connectionHops[0] && // ChanOpenInit will only set a single connection hop
 			previousChannel.Version == version) {
 			return "", nil, sdkerrors.Wrap(types.ErrInvalidChannel, "channel fields mismatch previous channel fields")
 		}
@@ -189,12 +174,7 @@ func (k Keeper) ChanOpenTry(
 	// NOTE: this step has been switched with the one below to reverse the connection
 	// hops
 	channel := types.NewChannel(types.TRYOPEN, order, counterparty, connectionHops, version)
-
-	counterpartyHops, found := k.CounterpartyHops(ctx, channel)
-	if !found {
-		// should not reach here, connectionEnd was able to be retrieved above
-		panic("cannot find connection")
-	}
+	counterpartyHops := []string{connectionEnd.GetCounterparty().GetConnectionID()}
 
 	// expectedCounterpaty is the counterparty of the counterparty's channel end
 	// (i.e self)
@@ -297,11 +277,7 @@ func (k Keeper) ChanOpenAck(
 		)
 	}
 
-	counterpartyHops, found := k.CounterpartyHops(ctx, channel)
-	if !found {
-		// should not reach here, connectionEnd was able to be retrieved above
-		panic("cannot find connection")
-	}
+	counterpartyHops := []string{connectionEnd.GetCounterparty().GetConnectionID()}
 
 	// counterparty of the counterparty channel end (i.e self)
 	expectedCounterparty := types.NewCounterparty(portID, channelID)
@@ -381,11 +357,7 @@ func (k Keeper) ChanOpenConfirm(
 		)
 	}
 
-	counterpartyHops, found := k.CounterpartyHops(ctx, channel)
-	if !found {
-		// Should not reach here, connectionEnd was able to be retrieved above
-		panic("cannot find connection")
-	}
+	counterpartyHops := []string{connectionEnd.GetCounterparty().GetConnectionID()}
 
 	counterparty := types.NewCounterparty(portID, channelID)
 	expectedChannel := types.NewChannel(
@@ -519,11 +491,7 @@ func (k Keeper) ChanCloseConfirm(
 		)
 	}
 
-	counterpartyHops, found := k.CounterpartyHops(ctx, channel)
-	if !found {
-		// Should not reach here, connectionEnd was able to be retrieved above
-		panic("cannot find connection")
-	}
+	counterpartyHops := []string{connectionEnd.GetCounterparty().GetConnectionID()}
 
 	counterparty := types.NewCounterparty(portID, channelID)
 	expectedChannel := types.NewChannel(